backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
scikit-learn
numpy
pdfplumber
python-docx
pydantic
openai
python-dotenv
